Move default specific to default impl
