+++ test/unit/lib/default/generateCommand.test.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const assert = chai.assert;
 
 const generateCommand = require('../../../../lib/default/generateCommand.js');
-const shared = require('../../../../lib/shared.js');
+const shared = require('../../../../lib/default/shared.js');
 
 describe('Gladys generate AwoX command', function () {
